% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2_Score.R
\name{R2_Score}
\alias{R2_Score}
\title{Compute Coefficient of Determination (R²)}
\usage{
R2_Score(y, yhat)
}
\arguments{
\item{y}{Numeric vector of true values.}

\item{yhat}{Numeric vector of predicted values.}
}
\value{
Numeric scalar R² score.
}
\description{
Compute Coefficient of Determination (R²)
}
