% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reasonbench-package.R
\docType{package}
\name{reasonbench-package}
\alias{reasonbench}
\alias{reasonbench-package}
\title{Reasonably Perfect Benchmark Framework}
\description{
Provides utilities for evaluating machine learning models with respect to realistic,
noise-aware benchmarks.
}
\author{
\strong{Maintainer}: Diogo Ribeiro \email{contact@diogoribeiro.ai}

}
\keyword{internal}
